Cleanup: move handlers to handlers.go
